Add token

+++ app/Http/Controllers/ApiTokenController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ApiTokenLoginRequest;
 use App\Http\Requests\ApiTokenRegisterRequest;
 use App\Models\User;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 
 class ApiTokenController extends Controller
@@ -23,17 +25,18 @@ class ApiTokenController extends Controller
         }
 
         $user = User::create([
+
             'name' => $request->name,
             'email' => $request->email,
-            'password' => Hash::make($request->password)
+            'password' => Hash::make($request->password),
         ]);
-
+        $user->created_at = now()->format('Y-m-d');
         $token = $user->createToken($request->token_name);
 
-        return [
+        return response()->json([
             'token' => $token->plainTextToken,
             'user' => $user
-        ];
+        ]);
 
     }
 
@@ -47,18 +50,21 @@ class ApiTokenController extends Controller
         $user->tokens()->where('name', $request->token_name)->delete();
 
         $token = $user->createToken($request->token_name);
-        // Abilities
-        //$token = $user->createToken($request->token_name, ['repo:view']);
-
-        return [
+        $response = response()->json([
             'token' => $token->plainTextToken,
             'user' => $user
-        ];
+        ]);
+
+        return $response->header('Content-Type', 'application/json');
+
     }
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        if ($request->user()) {
+            $request->user()->currentAccessToken()->delete();
+        }
+
 
         return response(null, 204);
     }
